+++ src/main/kotlin/ee/carlrobert/codegpt/completions/factory/CodeGPTRequestFactory.kt
@@ -10,7 +10,10 @@ import ee.carlrobert.codegpt.completions.ChatCompletionParameters
 import ee.carlrobert.codegpt.completions.factory.OpenAIRequestFactory.Companion.buildOpenAIMessages
 import ee.carlrobert.codegpt.psistructure.ClassStructureSerializer
 import ee.carlrobert.codegpt.settings.configuration.ConfigurationSettings
-import ee.carlrobert.codegpt.settings.service.codegpt.CodeGPTServiceSettings
+import ee.carlrobert.codegpt.settings.models.ModelSettings
+import ee.carlrobert.codegpt.settings.service.FeatureType
+import ee.carlrobert.codegpt.settings.service.ModelSelectionService
+import ee.carlrobert.codegpt.ui.textarea.ConversationTagProcessor
 import ee.carlrobert.codegpt.util.file.FileUtil
 import ee.carlrobert.llm.client.codegpt.request.chat.*
 import ee.carlrobert.llm.client.openai.completion.request.OpenAIChatCompletionStandardMessage
@@ -19,10 +22,13 @@ class CodeGPTRequestFactory(private val classStructureSerializer: ClassStructure
     BaseRequestFactory() {
 
     override fun createChatRequest(params: ChatCompletionParameters): ChatCompletionRequest {
-        val model = service<CodeGPTServiceSettings>().state.chatCompletionSettings.model
+        val model = ModelSelectionService.getInstance().getModelForFeature(FeatureType.CHAT)
+
         val configuration = service<ConfigurationSettings>().state
         val requestBuilder: ChatCompletionRequest.Builder =
-            ChatCompletionRequest.Builder(buildOpenAIMessages(model, params, emptyList()))
+            ChatCompletionRequest.Builder(
+                buildOpenAIMessages(model, params, emptyList(), emptyList())
+            )
                 .setModel(model)
                 .setSessionId(params.sessionId)
                 .setStream(true)
@@ -74,11 +80,15 @@ class CodeGPTRequestFactory(private val classStructureSerializer: ClassStructure
             )
         }.orEmpty()
 
-        val contextFilesWithPsi = contextFiles + psiContext
-        if (contextFilesWithPsi.isNotEmpty()) {
-            requestBuilder.setContext(AdditionalRequestContext(contextFilesWithPsi))
+        val conversationsHistory = params.history?.joinToString("\n\n") {
+            ConversationTagProcessor.formatConversation(it)
         }
-
+        requestBuilder.setContext(
+            AdditionalRequestContext(
+                contextFiles + psiContext,
+                conversationsHistory
+            )
+        )
         return requestBuilder.build()
     }
 
@@ -101,9 +111,10 @@ class CodeGPTRequestFactory(private val classStructureSerializer: ClassStructure
         systemPrompt: String,
         userPrompt: String,
         maxTokens: Int,
-        stream: Boolean
+        stream: Boolean,
+        featureType: FeatureType
     ): ChatCompletionRequest {
-        val model = service<CodeGPTServiceSettings>().state.chatCompletionSettings.model
+        val model = ModelSelectionService.getInstance().getModelForFeature(featureType)
         if (model == "o4-mini") {
             return buildBasicO1Request(model, userPrompt, systemPrompt, maxTokens, stream = stream)
         }