+++ src/main/kotlin/ee/carlrobert/codegpt/credentials/CredentialsStore.kt
@@ -61,16 +61,16 @@ object CredentialsStore {
             override val value: String = "ANTHROPIC_API_KEY"
         }
 
-        data object LlamaApiKey : CredentialKey() {
-            override val value: String = "LLAMA_API_KEY"
-        }
-
         data object GoogleApiKey : CredentialKey() {
             override val value: String = "GOOGLE_API_KEY"
         }
 
         data object OllamaApikey : CredentialKey() {
             override val value: String = "OLLAMA_API_KEY"
+        }
+
+        data object MistralApiKey : CredentialKey() {
+            override val value: String = "MISTRAL_API_KEY"
         }
     }
 }